refactor: move array creation routines out of lax_numpy.py
